Able to add/remove hardware from project
User can enter a a number of hardware units to add to their project.
Hardware allocated to project and project ID are displayed at the top.

Each button press subtracts amount of hardware units from hardware set
and adds that many units to the project.

A refresh button is used to view the current amount of hardware for the
project and hardware sets. MUST BE PRESSED at this point in order to
view any changes in hardware amount.